Clean up sort of

+++ src/main/java/ganymedes01/etfuturum/entities/EntityRabbit.java
@@ -113,7 +113,7 @@ public class EntityRabbit extends EntityAnimal {
 	@Override
 	protected void entityInit() {
 		super.entityInit();
-		dataWatcher.addObject(18, Byte.valueOf((byte) 0));
+		dataWatcher.addObject(18, (byte) 0);
 	}
 
 	@Override
